upd error

+++ src/main/java/in/codingboy/aioofbot/AioobbotApplication.java
@@ -35,7 +35,7 @@ public class AioobbotApplication {
                 URL url = new URL("https://xm1nbot.herokuapp.com/");
                 HttpURLConnection connection = (HttpURLConnection) url.openConnection();
                 connection.setRequestMethod("GET");
-                int responcode = connection.getResponseCode();
+                long responcode = connection.getResponseCode();
                 System.out.println("RS CODE:"+responcode);
                 if (responcode == HttpURLConnection.HTTP_OK)
                 {
